requirements.txt
apscheduler
django
gunicorn
django-heroku
requests
aylien-apiclient==0.7.0
djangorestframework
django-cors-headers==3.2.0
asyncio==3.4.3